@@ -13,19 +13,22 @@
     "@types/react-dom": "^17.0.11",
     "autoprefixer": "^10.4.2",
     "clsx": "^1.1.1",
+    "graphql": "^16.3.0",
     "postcss": "^8.4.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0-rc.0",
+    "react-dom": "^18.0.0-rc.0",
     "react-scripts": "5.0.0",
     "tailwindcss": "^3.0.18",
     "typescript": "^4.5.5",
+    "urql": "^2.1.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "dev": "PORT=8000 craco start",
     "build": "craco build",
     "test": "craco test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "generate": "graphql-codegen --config codegen.yml"
   },
   "eslintConfig": {
     "extends": [
@@ -46,6 +49,13 @@
     ]
   },
   "devDependencies": {
-    "@types/webpack-env": "^1.16.3"
+    "@types/webpack-env": "^1.16.3",
+    "@graphql-codegen/add": "^3.1.1",
+    "@graphql-codegen/cli": "2.4.0",
+    "@graphql-codegen/introspection": "2.1.1",
+    "@graphql-codegen/typescript": "2.4.2",
+    "@graphql-codegen/typescript-operations": "2.2.3",
+    "@graphql-codegen/typescript-urql": "^3.5.0",
+    "@graphql-codegen/urql-introspection": "2.1.1"
   }
 }
