@@ -1,5 +1,15 @@
-import * as React from "react";
-import Text from "@components/Text";
+import { Suspense } from "react";
+
+import { Text } from "@components/Text";
+import { useCheckoutQuery } from "@graphql";
+
+const SuspenseTest = () => {
+  const [{ data }] = useCheckoutQuery({
+    variables: { token: "fc03e431-ff48-4e7f-b953-7ad6987bbcbc" },
+  });
+
+  return <span>{data?.checkout?.email}</span>;
+};
 
 export const Checkout = () => {
   return (
@@ -8,7 +18,9 @@ export const Checkout = () => {
         Saleor Checkout
       </Text>
       <Text size="sm" color="secondary">
-        (Greatest checkout ever)
+        <Suspense fallback={"Loading..."}>
+          <SuspenseTest />
+        </Suspense>
       </Text>
     </div>
   );
